requirements.txt
# Core
fastapi>=0.109.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Async HTTP
httpx>=0.25.0
aiofiles>=23.2.0

# LLM Client (OpenAI-compatible for Qwen)
openai>=1.3.0

# Voyage AI Embeddings (replaces sentence-transformers)
# Note: We use httpx directly for API calls, no SDK needed
# voyageai>=0.2.0  # Optional: official SDK

# Async utilities
nest-asyncio>=1.5.0

# Vector Database
qdrant-client>=1.7.0

# Image Processing
Pillow>=10.3.0
PyMuPDF>=1.23.26  # For PDF extraction

# Object Storage (optional)
boto3>=1.33.0  # For S3
# oss2>=2.18.0  # For Alibaba OSS (uncomment if needed)

# Utilities
python-multipart>=0.0.9  # For file uploads
numpy>=1.24.0
